+++ src/editor/chip-view.ts
@@ -81,6 +81,10 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
 
         webviewView.webview.onDidReceiveMessage((message) => {
             if (message.type === "selectPin") {
+                if(message.pin === undefined || message.pin === null){
+                    console.log("Pin not passed in message.");
+                    return;
+                }
                 console.log(`[Chip View] selected pin ` + message.pin.id);
                 providerPinView.selectPin(message.pin.id);
                 this.selectPin(Object.assign({pin: message.pin.id, pinType: message.pin.type}));
@@ -112,7 +116,9 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
             this._view.show?.(true);
             //get pin declarations. If pld file opened, get from editor window, otherwise get from disk
             //const pldFile = vscode.workspace.textDocuments.some(doc => doc.uri.toString() === uri.toString());
+            
             if(vscode.window.activeTextEditor?.document === undefined){
+                this._view.webview.postMessage({ type: "selectPin", pin: pin });
                 return;
             }
             const pinDeclarations = vscode.window.activeTextEditor?.document
@@ -298,7 +304,7 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
                 });
                 providerChipView._colors.push({
                     type: "accent1",
-                    color: "#eef",
+                    color: "#445",
                 });
                 providerChipView._colors.push({
                     type: "accent2",
@@ -407,11 +413,11 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
                 });
                 providerChipView._colors.push({
                     type: "pinIN",
-                    color: "#ACEDAF",
+                    color: "#90a5e9ff",
                 });
                 providerChipView._colors.push({
                     type: "pinINOUT",
-                    color: "#adFFEF",
+                    color: "#69bccaff",
                 });
                 providerChipView._colors.push({
                     type: "pinOUT",
@@ -419,7 +425,7 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
                 });
                 providerChipView._colors.push({
                     type: "pinOE",
-                    color: "#C9C596",
+                    color: "#c9bc2bff",
                 });
                 providerChipView._colors.push({
                     type: "pinCLR",
@@ -538,7 +544,7 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
 
     public openProjectChipView(project: Project | undefined) {
         if (project === undefined) {
-            extensionState.setActiveProject(undefined);
+            //extensionState.setActiveProject(undefined);
             this.setDevice(undefined);
             //this.setColors();
             providerPinView.setPins(undefined);
@@ -547,7 +553,7 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
         if (project.devicePins) {
             const pins = project.devicePins;
             if (pins) {
-                extensionState.setActiveProject(project);
+                //extensionState.setActiveProject(project);
                 this.setDevice(pins);
                 this.setColors();
                 providerPinView.setPins(pins);
@@ -680,7 +686,7 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
     private _getHtmlForWebview(webview: vscode.Webview) {
         // Get the local path to main script run in the webview, then convert it to a uri we can use in the webview.
         const scriptUri = webview.asWebviewUri(
-            vscode.Uri.joinPath(this._extensionUri, "assets", "js", "script.js")
+            vscode.Uri.joinPath(this._extensionUri, "assets", "js", "chipView.js")
         );
 
         // Do the same for the stylesheet.
@@ -735,8 +741,7 @@ export class ChipViewProvider implements vscode.WebviewViewProvider {
 				<title>Chip View</title>
 			</head>
 			<body>
-				<canvas id='ic' width="800" height="500"></canvas>
-
+				<div id="chip"></div>
 				
 				<script nonce="${nonce}" src="${scriptUri}"></script>				
 			</body>